Trim Blocks\Setup to registration/wiring methods only.
