+++ includes/core/classes/blocks/class-dropdown.php
@@ -104,7 +104,6 @@ class Dropdown {
 	 * @return string The modified block content with inline styles.
 	 */
 	public function generate_block_styles( string $block_content, array $block ): string {
-		$block_instance         = Setup::get_instance();
 		$attributes             = $block['attrs'] ?? array();
 		$dropdown_id            = $attributes['dropdownId'] ?? '';
 		$open_on                = $attributes['openOn'] ?? 'click';
@@ -159,7 +158,7 @@ class Dropdown {
 						display: block;
 					}
 				',
-				esc_attr( $block_instance->get_default_block_class( $block['blockName'] ) ),
+				esc_attr( wp_get_block_default_classname( $block['blockName'] ) ),
 				esc_attr( $dropdown_id )
 			);
 		}
