+++ test/unit/php/includes/tests/core/classes/blocks/class-test-setup.php
@@ -113,130 +113,6 @@ class Test_Setup extends Base {
 	}
 
 	/**
-	 * Coverage for get_block_variations.
-	 *
-	 * @covers ::get_block_variations
-	 *
-	 * @return void
-	 */
-	public function test_get_block_variations(): void {
-		$instance = Setup::get_instance();
-
-		$this->assertSame(
-			array(
-				'query',
-				'query-no-results',
-				'query-pagination',
-				'query-pagination-next',
-				'query-pagination-numbers',
-				'query-pagination-previous',
-			),
-			$instance->get_block_variations(),
-			'Failed to assert, to get all block variations from the "/src" directory.'
-		);
-	}
-
-	/**
-	 * Coverage for get_block_variations when directory doesn't exist.
-	 *
-	 * Covers: Early return when variations directory doesn't exist.
-	 *
-	 * @covers ::get_block_variations
-	 *
-	 * @return void
-	 */
-	public function test_get_block_variations_directory_not_exists(): void {
-		$instance            = Setup::get_instance();
-		$variations_dir      = sprintf( '%1$s/build/variations/core/', GATHERPRESS_CORE_PATH );
-		$temp_renamed_dir    = sprintf( '%1$s/build/variations/core-temp-renamed/', GATHERPRESS_CORE_PATH );
-		$variations_dir_base = sprintf( '%1$s/build/variations/', GATHERPRESS_CORE_PATH );
-
-		// Temporarily rename the variations directory to simulate non-existence.
-		if ( file_exists( $variations_dir ) ) {
-			// phpcs:ignore WordPress.WP.AlternativeFunctions.rename_rename -- Necessary for testing.
-			rename( $variations_dir, $temp_renamed_dir );
-		}
-
-		// Reset the cached property to force a fresh check.
-		Utility::set_and_get_hidden_property( $instance, 'block_variation_names', array() );
-
-		// Now the directory doesn't exist, should return empty array.
-		$result = $instance->get_block_variations();
-
-		$this->assertSame( array(), $result, 'Should return empty array when variations directory does not exist.' );
-
-		// Restore the directory.
-		if ( file_exists( $temp_renamed_dir ) ) {
-			// phpcs:ignore WordPress.WP.AlternativeFunctions.rename_rename -- Necessary for testing.
-			rename( $temp_renamed_dir, $variations_dir );
-		}
-
-		// Reset the cache again for other tests.
-		Utility::set_and_get_hidden_property( $instance, 'block_variation_names', array() );
-	}
-
-	/**
-	 * Coverage for get_block_variations caching behavior.
-	 *
-	 * Covers: Caching of block variation names.
-	 *
-	 * @covers ::get_block_variations
-	 *
-	 * @return void
-	 */
-	public function test_get_block_variations_caching(): void {
-		$instance = Setup::get_instance();
-
-		// Reset the cache to ensure we're starting fresh.
-		Utility::set_and_get_hidden_property( $instance, 'block_variation_names', array() );
-
-		// Verify block_variation_names is empty initially.
-		$cache_before = Utility::get_hidden_property( $instance, 'block_variation_names' );
-		$this->assertEmpty( $cache_before );
-
-		// First call should populate the cache.
-		$first_result = $instance->get_block_variations();
-
-		// Verify cache is now populated (target code executed).
-		$cache_after_first = Utility::get_hidden_property( $instance, 'block_variation_names' );
-		$this->assertNotEmpty( $cache_after_first );
-
-		// Second call should use cached values (target code check causes early return from cache).
-		$second_result = $instance->get_block_variations();
-
-		// Verify cache wasn't modified by second call.
-		$cache_after_second = Utility::get_hidden_property( $instance, 'block_variation_names' );
-		$this->assertSame( $cache_after_first, $cache_after_second );
-
-		// Both results should be identical.
-		$this->assertSame( $first_result, $second_result, 'Should return cached variation names on subsequent calls.' );
-
-		// Verify the final result matches the cached data after array_filter.
-		$this->assertSame( array_filter( $cache_after_second ), $second_result );
-	}
-
-	/**
-	 * Coverage for get_classname_from_foldername.
-	 *
-	 * @covers ::get_classname_from_foldername
-	 *
-	 * @return void
-	 */
-	public function test_get_classname_from_foldername(): void {
-		$instance = Setup::get_instance();
-
-		$this->assertSame(
-			'Unit_Test',
-			Utility::invoke_hidden_method(
-				$instance,
-				'get_classname_from_foldername',
-				array( '/src/variations/unit-test' )
-			),
-			'Failed to assert, to get class name from foldername.'
-		);
-	}
-
-	/**
 	 * Coverage for register_block_patterns.
 	 *
 	 * @covers ::register_block_patterns
@@ -301,35 +177,6 @@ class Test_Setup extends Base {
 		$instance->register_block_classes();
 
 		$this->assertTrue( true );
-	}
-
-	/**
-	 * Coverage for get_default_block_class.
-	 *
-	 * @covers ::get_default_block_class
-	 *
-	 * @return void
-	 */
-	public function test_get_default_block_class(): void {
-		$instance = Setup::get_instance();
-
-		$this->assertSame(
-			'wp-block-gatherpress-event-date',
-			$instance->get_default_block_class( 'gatherpress/event-date' ),
-			'Failed to generate correct block class for gatherpress/event-date.'
-		);
-
-		$this->assertSame(
-			'wp-block-core-paragraph',
-			$instance->get_default_block_class( 'core/paragraph' ),
-			'Failed to generate correct block class for core/paragraph.'
-		);
-
-		$this->assertSame(
-			'wp-block-my-plugin-custom-block',
-			$instance->get_default_block_class( 'my-plugin/custom-block' ),
-			'Failed to generate correct block class for custom block.'
-		);
 	}
 
 	/**
@@ -768,111 +615,6 @@ class Test_Setup extends Base {
 
 		// Should return the original block unchanged (no placeholder added for non-paragraph).
 		$this->assertSame( $parsed_hooked_block, $result );
-	}
-
-	/**
-	 * Coverage for get_block_names with simple block.
-	 *
-	 * @covers ::get_block_names
-	 *
-	 * @return void
-	 */
-	public function test_get_block_names_simple(): void {
-		$instance = Setup::get_instance();
-
-		$blocks = array(
-			'blockName' => 'core/paragraph',
-		);
-
-		$result = $instance->get_block_names( $blocks );
-
-		$this->assertSame( array( 'core/paragraph' ), $result );
-	}
-
-	/**
-	 * Coverage for get_block_names with nested blocks.
-	 *
-	 * @covers ::get_block_names
-	 *
-	 * @return void
-	 */
-	public function test_get_block_names_nested(): void {
-		$instance = Setup::get_instance();
-
-		$blocks = array(
-			'blockName'   => 'core/group',
-			'innerBlocks' => array(
-				array(
-					'blockName' => 'core/paragraph',
-				),
-				array(
-					'blockName' => 'gatherpress/event-date',
-				),
-			),
-		);
-
-		$result = $instance->get_block_names( $blocks );
-
-		$this->assertSame(
-			array( 'core/group', 'core/paragraph', 'gatherpress/event-date' ),
-			$result
-		);
-	}
-
-	/**
-	 * Coverage for get_block_names with deeply nested blocks.
-	 *
-	 * @covers ::get_block_names
-	 *
-	 * @return void
-	 */
-	public function test_get_block_names_deeply_nested(): void {
-		$instance = Setup::get_instance();
-
-		$blocks = array(
-			'blockName'   => 'core/group',
-			'innerBlocks' => array(
-				array(
-					'blockName'   => 'core/columns',
-					'innerBlocks' => array(
-						array(
-							'blockName'   => 'core/column',
-							'innerBlocks' => array(
-								array(
-									'blockName' => 'core/paragraph',
-								),
-							),
-						),
-					),
-				),
-			),
-		);
-
-		$result = $instance->get_block_names( $blocks );
-
-		$this->assertSame(
-			array( 'core/group', 'core/columns', 'core/column', 'core/paragraph' ),
-			$result
-		);
-	}
-
-	/**
-	 * Coverage for get_block_names with no blockName.
-	 *
-	 * @covers ::get_block_names
-	 *
-	 * @return void
-	 */
-	public function test_get_block_names_no_blockname(): void {
-		$instance = Setup::get_instance();
-
-		$blocks = array(
-			'attrs' => array(),
-		);
-
-		$result = $instance->get_block_names( $blocks );
-
-		$this->assertSame( array(), $result );
 	}
 
 	/**
